requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1

# Spotify API
spotipy>=2.23.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=7.0.0

# Development tools
black>=23.0.0
pylint>=2.17.0
